When adding a card, prefill form based on current View
Fixes #287

- Adds a new base view, used for discussing Jellyfish
- Threads can now be tagged, allowing for the creation of 'rooms'

change-type: minor
