+++ default-cards/contrib/action-create-card.json
@@ -40,7 +40,7 @@
             "type": "array",
             "items": {
               "type": "string",
-              "pattern": "^[a-zA-Z0-9-]+$"
+              "pattern": "^[a-zA-Z0-9-/]+$"
             }
           },
           "links": {
