+++ default-cards/contrib/action-update-card.json
@@ -25,7 +25,7 @@
             "type": "array",
             "items": {
               "type": "string",
-              "pattern": "^[a-zA-Z0-9-]+$"
+              "pattern": "^[a-zA-Z0-9-/]+$"
             }
           },
           "links": {
