+++ default-cards/contrib/action-upsert-card.json
@@ -36,7 +36,7 @@
             "type": "array",
             "items": {
               "type": "string",
-              "pattern": "^[a-zA-Z0-9-]+$"
+              "pattern": "^[a-zA-Z0-9-/]+$"
             }
           },
           "links": {
