+++ default-cards/contrib/thread.json
@@ -9,6 +9,13 @@
     "schema": {
       "type": "object",
       "properties": {
+        "tags": {
+          "type": "array",
+          "items": {
+            "type": "string",
+            "pattern": "^[a-zA-Z0-9-/]+$"
+          }
+        },
         "data": {
           "type": "object",
           "properties": {
