+++ default-cards/contrib/update.json
@@ -39,7 +39,7 @@
                   "type": "array",
                   "items": {
                     "type": "string",
-                    "pattern": "^[a-zA-Z0-9-]+$"
+                    "pattern": "^[a-zA-Z0-9-/]+$"
                   }
                 },
                 "links": {
