+++ default-cards/contrib/view-my-alerts.json
@@ -30,7 +30,7 @@
       }
     ],
     "lenses": [
-      "lens-my-mentions"
+      "lens-interleaved"
     ]
   }
 }
