+++ default-cards/contrib/view-my-mentions.json
@@ -30,7 +30,7 @@
       }
     ],
     "lenses": [
-      "lens-my-mentions"
+      "lens-interleaved"
     ]
   }
 }
