+++ default-cards/core/card.json
@@ -29,7 +29,7 @@
           "type": "array",
           "items": {
             "type": "string",
-            "pattern": "^[a-zA-Z0-9-]+$"
+            "pattern": "^[a-zA-Z0-9-/]+$"
           }
         },
         "links": {
