+++ lib/server.js
@@ -155,6 +155,7 @@ const createServer = () => {
 			require('../default-cards/contrib/view-my-mentions.json'),
 			require('../default-cards/contrib/view-my-alerts.json'),
 			require('../default-cards/contrib/view-non-executed-action-requests.json'),
+			require('../default-cards/contrib/view-p-jellyfish.json'),
 			require('../default-cards/contrib/view-read-user-community.json'),
 			require('../default-cards/contrib/view-read-user-guest.json'),
 			require('../default-cards/contrib/view-read-user-team-admin.json'),
