+++ lib/ui/services/sdk-helpers.ts
@@ -1,0 +1,19 @@
+import { JSONSchema6 } from 'json-schema';
+import * as _ from 'lodash';
+import { Card } from '../../Types';
+import { sdk } from '../app';
+import { getViewSchema } from './helpers';
+
+export const loadSchema = async (query: string | Card | JSONSchema6) => {
+	if (_.isString(query)) {
+		return await sdk.card.get(query).toPromise()
+			.then(getViewSchema);
+	}
+
+	if (query.type === 'view') {
+		return getViewSchema(query as Card);
+	}
+
+	return query as JSONSchema6;
+};
+
