+++ lib/ui/services/subscriptions.ts
@@ -2,7 +2,7 @@ import * as _ from 'lodash';
 import { JellyfishStream } from '../../sdk/stream';
 import { Card } from '../../Types';
 import { actionCreators, sdk, store } from '../app';
-import { loadSchema } from '../services/helpers';
+import { loadSchema } from '../services/sdk-helpers';
 import { createNotification } from './notifications';
 
 export class SubscriptionManager {
