+++ package.json
@@ -122,7 +122,7 @@
     "react-trello": "^1.31.3",
     "redux": "^3.7.2",
     "redux-thunk": "^2.2.0",
-    "rendition": "4.8.1",
+    "rendition": "4.9.0",
     "rethinkdb": "^2.3.3",
     "rxjs": "^6.2.0",
     "socket.io": "^2.1.0",
